Rename narrowSelectors to narrowsSelectors.
